[Node] Redesign node web interface settings page and split into subcomponents
Merges node type and sync host into a single "Node configuration" card,
reworks the Support card into a "Help" card with Guides/Support/Developer
sections, relocates the stop-node and log-out actions into a shared
CardCornerButton, and switches wallet/encryption-key layouts to a denser
grid. Splits the 1350+ line settings.tsx into per-card components under
components/Settings/ to keep files small and focused.
